requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ale-py==0.8.1
altgraph==0.17.4
amqp==5.2.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.11.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.2
billiard==4.2.1
bleach==6.0.0
blinker==1.9.0
breadability==0.1.20
bs4==0.0.2
build==1.3.0
cachetools==5.3.2
camelot-py==0.9.0
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
configparser==3.5.3
contourpy==1.1.1
cryptography==41.0.7
cssselect==1.2.0
cupy-cuda11x==13.0.0
cycler==0.12.1
data==0.4
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
docopt==0.6.2
durationpy==0.10
Elixir==0.7.1
et-xmlfile==1.1.0
excalibur-py==0.4.3
executing==1.2.0
fake-useragent==1.4.0
Farama-Notifications==0.0.4
fastapi==0.124.0
fastjsonschema==2.18.0
fastrlock==0.8.2
filelock==3.13.3
Flask==3.1.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.8.0
fsspec==2024.3.1
funcsigs==1.0.2
future==1.0.0
gast==0.5.4
ghostscript==0.7
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.76.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.7.1
httpx==0.26.0
huggingface-hub==0.29.1
humanfriendly==10.0
ibm-cloud-sdk-core==3.18.2
ibm-platform-services==0.49.0
idna==3.4
imageio==2.37.0
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
jiter==0.12.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.15.0
kiwisolver==1.4.5
kombu==5.4.2
kubernetes==34.1.0
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.147
latex==0.7.0
lazy_loader==0.4
libclang==16.0.6
llvmlite==0.45.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==4.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
MetaTrader5==5.0.4200
mistune==3.0.1
ml-dtypes==0.2.0
mmh3==5.2.0
mpl-finance==0.10.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.11
mypy_extensions==1.1.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.9.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.62.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.9.0.80
opencv-python-headless==4.12.0.88
openpyxl==3.1.2
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-util-http==0.60b0
opt-einsum==3.3.0
orjson==3.11.5
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parse==1.20.1
parso==0.8.3
pbr==6.0.0
pdfminer.six==20240706
peewee==3.17.1
pefile==2023.2.7
pickleshare==0.7.5
pillow==11.3.0
platformdirs==3.10.0
ply==3.11
pooch==1.8.2
posthog==7.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==5.9.5
pulsar-client==3.8.0
pure-eval==0.2.2
py2exe==0.13.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==24.6.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pyee==8.2.2
pyfiglet==1.0.2
pygame==2.5.2
Pygments==2.16.1
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.8.0
pylatexenc==2.10
PyMatting==1.1.14
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==2.12.1
PyPika==0.48.9
pyppeteer==1.0.2
pyproject_hooks==1.2.0
pyquery==2.0.0
pyreadline3==3.5.4
PySocks==1.7.1
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qiskit==0.45.1
qiskit-aer==0.13.1
qiskit-ibm-provider==0.8.0
qiskit-ibm-runtime==0.17.0
qiskit-terra==0.45.1
referencing==0.30.2
regex==2024.11.6
rembg==2.0.67
requests==2.31.0
requests-file==2.0.0
requests-html==0.10.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.10.3
rsa==4.9
rustworkx==0.13.2
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.3.1
scipy==1.16.2
seaborn==0.13.0
selenium==4.17.2
Send2Trash==1.8.2
shellingham==1.5.4
Shimmy==0.2.1
shutilwhich==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
sqlalchemy-migrate==0.11.0
sqlparse==0.5.1
sspilib==0.1.0
stack-data==0.6.2
starlette==0.50.0
stevedore==5.1.0
sumy==0.11.0
sympy==1.12
tabula==1.0.5
tabula-py==2.9.3
tabulate==0.9.0
tempdir==0.7.1
Tempita==0.5.2
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2025.9.20
tiktoken==0.12.0
tinycss2==1.2.1
tldextract==5.1.2
tokenizers==0.21.0
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.49.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.20.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.38.0
validators==0.28.1
vine==5.1.0
w3lib==2.1.2
watchfiles==1.1.1
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.3
websockets==10.4
Werkzeug==3.1.3
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xlrd==0.7.1
xlwt==0.7.2
yarl==1.22.0
yfinance==0.2.36
zipp==3.17.0
